Revamp error handling
This change uses the `failure` crate as a basis for handling errors in
the codebase. Pulling in `failure` reduced the variants on the `Error`
type and also allowed the `pretty_error` function to be useful. The
approach is using `failure`'s `ResultExt` trait heavily to add context
to the errors in an attempt to make the human error printing more
relevant.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>