Use upstream human-panic 1.0.1 for Crate release
This change reverts the codebase to use the vanilla upstream version of
`human-panic` so this tool can be published on Crates.io in the
meantime. Unfortuntely, this adds multiple crates which are **not** used
and are purely extra.

When some of the upstream PRs and issues are resolved, merged, and
released, then this commit may become revertable.

References: rust-cli/human-panic#46

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>